Fix not primary auto_increment case
The assumption that auto_increment is always the value for primary key is wrong.

As example:
* there is a table user, where primary id is - userId which is a md5 random string.
* Autoincrement is on the userNumber field, which is not primary, but by some reason it is there.

Now we add a user with haveInDatabase, and with specific id:['userId'=>'special'], what will endup in insertedRows? usesrId => 10 for example, as insertId for user was 10, and sure, it will not be cleaned up.